requirements.txt
asgiref==3.4.1
boto3==1.18.44
botocore==1.21.44
certifi==2021.5.30
Django==3.2.7
djangorestframework==3.12.4
jmespath==0.10.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
